Okay, now only need to finish the implementation for online games
